singularize store
