+++ expo-zustand-styled-components/src/hooks/stores/index.ts
@@ -1,3 +1,2 @@
 export { default as useAuthStore } from './useAuthStore';
-export { default as useGistsStore } from './useGistsStore';
-export { default as useRepoStore } from './useRepoStore';
+export { default as useAppStore } from './useAppStore';
