+++ expo-zustand-styled-components/src/services/get-gists.ts
@@ -1,6 +1,7 @@
 import { USER_GISTS_QUERY } from './queries/gists';
-import FetchApi, { ApiProps } from './api';
+import FetchApi from './api';
 import { Gists } from '../types/gists-type';
+import { useAppStore } from '../hooks/stores';
 
 interface Response {
   data: {
@@ -13,39 +14,38 @@ interface Response {
 }
 
 const getGists = async () => {
-  const data: ApiProps<undefined> = {
-    query: USER_GISTS_QUERY,
-  };
-
-  const resp = (await FetchApi(data)) as Response;
-
-  const gists = resp.data?.viewer.gists.nodes?.reduce(
-    (acc: Gists[] | { name: string; url: string }[], gist) => {
-      if (!gist) {
-        return acc;
-      }
-      const files = gist.files ?? [];
-      const gists = files.reduce(
-        (_acc: { name: string; url: string }[], file) =>
-          file
-            ? [
-                ..._acc,
-                {
-                  id: gist.id,
-                  description: gist.description,
-                  name: file.name || gist.name,
-                  url: gist.url,
-                },
-              ]
-            : acc,
-        []
-      );
-      return [...acc, ...gists];
-    },
-    []
-  );
-
-  return gists;
+  try {
+    useAppStore.setState({ isLoading: true });
+    const resp = (await FetchApi({ query: USER_GISTS_QUERY })) as Response;
+    const gists = resp.data?.viewer.gists.nodes?.reduce(
+      (acc: Gists[] | { name: string; url: string }[], gist) => {
+        if (!gist) {
+          return acc;
+        }
+        const files = gist.files ?? [];
+        const gists = files.reduce(
+          (_acc: { name: string; url: string }[], file) =>
+            file
+              ? [
+                  ..._acc,
+                  {
+                    id: gist.id,
+                    description: gist.description,
+                    name: file.name || gist.name,
+                    url: gist.url,
+                  },
+                ]
+              : acc,
+          []
+        );
+        return [...acc, ...gists];
+      },
+      []
+    );
+    useAppStore.setState({ isLoading: false, gists });
+  } catch (err) {
+    useAppStore.setState({ isLoading: false, error: err.message });
+  }
 };
 
 export default getGists;
