+++ expo-zustand-styled-components/src/services/get-repo-file.ts
@@ -1,7 +1,7 @@
 import FetchApi from './api';
 import { REPO_FILE_QUERY } from './queries/repo-file';
-import { ApiProps } from './api';
 import { RepoFile } from '../types/repo-file-type';
+import { useAppStore } from '../hooks/stores';
 
 type RepoFileVariables = {
   owner: string;
@@ -15,30 +15,14 @@ type Response = {
   };
 };
 
-/**
- *
- * @param {
- *  variable: {
- *    owner
- *    name
- *    expression
- *  }
- * }
- */
-
 const getRepoFile = async (variables: RepoFileVariables) => {
-
-  const data: ApiProps<RepoFileVariables> = {
-    url: ``, // Missing url
-    query: REPO_FILE_QUERY,
-    variables,
-    headersOptions: {
-      authorization: `Bearer `, // Missing token
-    },
-  };
-  const resp = (await FetchApi(data)) as Response;
-
-  return resp?.data?.repository?.blob;
+  try {
+    useAppStore.setState({ isLoading: true });
+    const resp = (await FetchApi({ query: REPO_FILE_QUERY, variables })) as Response;
+    useAppStore.setState({ isLoading: false, file: resp?.data?.repository?.blob });
+  } catch (err) {
+    useAppStore.setState({ isLoading: false, error: err.message });
+  }
 };
 
 export default getRepoFile;
