+++ expo-zustand-styled-components/src/services/get-repo-info.ts
@@ -1,6 +1,7 @@
 import FetchApi from './api';
 import { REPO_INFO_QUERY } from './queries/repo-info';
-import { Info, RepoInfo, Topics, RepoInfoVariables } from '../types/repo-info-type';
+import { RepoInfo, Topics, RepoInfoVariables } from '../types/repo-info-type';
+import { useAppStore } from '../hooks/stores';
 
 type Response = {
   data: {
@@ -21,17 +22,12 @@ export function parseTopics(topics: Topics[]) {
   }, [] as string[]);
 }
 
-const getRepoInfo = async (
-  variables: RepoInfoVariables
-): Promise<{
-  branch: string;
-  info: Info;
-}> => {
-  const resp = (await FetchApi({ query: REPO_INFO_QUERY, variables })) as Response;
-  const repository = resp.data?.repository;
-  return {
-    branch: repository?.defaultBranchRef?.name ?? 'HEAD',
-    info: {
+const getRepoInfo = async (variables: RepoInfoVariables) => {
+  try {
+    useAppStore.setState({ isLoading: true });
+    const resp = (await FetchApi({ query: REPO_INFO_QUERY, variables })) as Response;
+    const repository = resp.data?.repository;
+    const info = {
       isPrivate: repository?.isPrivate,
       visibility: repository?.visibility,
       forkCount: repository?.forkCount,
@@ -43,8 +39,16 @@ const getRepoInfo = async (
       topics: parseTopics(repository?.topics?.nodes),
       isOrg: typeof repository?.owner?.orgName === 'string',
       openPullRequestCount: repository?.pullRequests?.totalCount,
-    },
-  };
+    };
+
+    useAppStore.setState({
+      isLoading: false,
+      info,
+      branch: repository?.defaultBranchRef?.name ?? 'HEAD',
+    });
+  } catch (err) {
+    useAppStore.setState({ isLoading: false, error: err.message });
+  }
 };
 
 export default getRepoInfo;
