+++ expo-zustand-styled-components/src/services/get-repo-readme.ts
@@ -1,6 +1,7 @@
 import FetchApi from './api';
 import { REPO_README_QUERY } from './queries/repo-readme';
 import { RepoReadme, RepoReadmeVariables } from '../types/repo-readme-type';
+import { useAppStore } from '../hooks/stores';
 
 type Response = {
   data: {
@@ -9,9 +10,13 @@ type Response = {
 };
 
 const getRepoReadMe = async (variables: RepoReadmeVariables) => {
-  const resp = (await FetchApi({ query: REPO_README_QUERY, variables })) as Response;
-
-  return resp.data.repository?.readme?.text;
+  try {
+    useAppStore.setState({ isLoading: true });
+    const resp = (await FetchApi({ query: REPO_README_QUERY, variables })) as Response;
+    useAppStore.setState({ isLoading: false, readMe: resp.data.repository?.readme?.text });
+  } catch (err) {
+    useAppStore.setState({ isLoading: false, error: err.message });
+  }
 };
 
 export default getRepoReadMe;
