+++ expo-zustand-styled-components/src/services/get-repo-tree.ts
@@ -1,6 +1,7 @@
 import FetchApi from './api';
 import { REPO_TREE_QUERY } from './queries/repo-tree';
 import { RepoTree, RepoTreeVariables } from '../types/repo-tree-type';
+import { useAppStore } from '../hooks/stores';
 
 type Response = {
   data: {
@@ -9,12 +10,17 @@ type Response = {
 };
 
 const getRepoTree = async (variables: RepoTreeVariables) => {
-  const resp = (await FetchApi({ query: REPO_TREE_QUERY, variables })) as Response;
-
-  return {
-    branches: resp.data?.repository?.branches?.nodes,
-    tree: resp.data?.repository?.tree?.entries,
-  };
+  try {
+    useAppStore.setState({ isLoading: true });
+    const resp = (await FetchApi({ query: REPO_TREE_QUERY, variables })) as Response;
+    useAppStore.setState({
+      isLoading: false,
+      tree: resp.data?.repository?.tree?.entries,
+      branches: resp.data?.repository?.branches?.nodes,
+    });
+  } catch (err) {
+    useAppStore.setState({ isLoading: false, error: err.message });
+  }
 };
 
 export default getRepoTree;
